+++ Qualifier.Api/Core/Qualifier.Application/Database/ControlGroup/Queries/GetAllControlGroupsByStandardId/GetAllControlGroupsByStandardIdDto.cs
@@ -3,7 +3,7 @@ namespace Qualifier.Application.Database.ControlGroup.Queries.GetAllControlGroup
     public class GetAllControlGroupsByStandardIdDto
     {
         public int controlGroupId { get; set; }
-        public int number { get; set; }
+        public decimal number { get; set; }
         public string name { get; set; }
 
     }
