+++ Qualifier.Api/Core/Qualifier.Application/Database/ControlGroup/Queries/GetControlGroupsByCompanyId/GetControlGroupsByCompanyIdDto.cs
@@ -3,7 +3,7 @@ namespace Qualifier.Application.Database.ControlGroup.Queries.GetControlGroupsBy
     public class GetControlGroupsByCompanyIdDto
     {
         public int controlGroupId { get; set; }
-        public int number { get; set; }
+        public decimal number { get; set; }
         public string name { get; set; }
 
     }
